requirements.txt
pygame==2.6.1
PyOpenGL==3.1.10
PyOpenGL-accelerate==3.1.10
numpy
